Port boot to use gnu assembler (gas).
I used the disassembled boot.o.txt to guide my translation to
boot.gas.S and tried to keep the generated output the same.
The only known instruction I couldn't translate was the long
jump to kmain and used .byte, .long, .word to fake it.

There are probably other issues.