Fix: Determine gnu assembler long jump instruction #1
The AT&T syntax I needed for ljmp was:

  ljmp  $gdt64_code_seg, $kmain

So the both the segment (gdt64_code_seg) and offset (kmain) needed to
be constants hence the '$'. Also, $gdt64_code_seg needed to be defined
before its use as I got a:

  "Error: can't handle non absolute segment in `ljmp'"

So I moved the Global Descriptor Table rodata section above the text
section for the code.