Resolve issue #2, Interrupts handlers were not being invoked.
The problem was that in set_idtr the compiler was optimizing away
the dp.itg = &idt[0]. The solution I have is to add volatile to the
limit and union in descriptor_ptr.

Another solution is to not add the u16 unused[3] but then we'll have
alignment problems in the future.

Also add other code while debugging, such ad expt_xx_handlers.

There is still a problem that the IRET is failing and and invokes
a expt_13_handler which is a General Protection fault.